Navbar Correction Done
