+++ src/app/contact-us/page.tsx
@@ -99,7 +99,7 @@ export default function ContactUs() {
     return (
         <main className="w-full max-w-[390px] md:max-w-[1512px] mx-auto min-h-screen bg-white text-black">
             {/* Banner Section */}
-            <div className="w-full bg-[#F4F3FF] pt-[65px]">
+            <div className="w-full bg-[#F4F3FF] pt-[70px]">
                 <div className="w-full max-w-[360px] md:max-w-[1280px] mx-auto py-12 md:py-16 lg:py-20 px-4 sm:px-6 lg:px-8 text-center">
                     <p className="text-[14px] md:text-[16px] font-semibold text-[#7856FC] mb-[16px] font-inter leading-[20px] md:leading-[24px] text-center">Contact Us</p>
                     <h1 className="text-[36px] md:text-[48px] font-semibold mb-4 font-inter leading-[44px] md:leading-[60px] tracking-[-0.02em] text-center">
